backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
asyncpg==0.29.0
alembic==1.13.0

# Pydantic and Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI and MCP
openai==1.12.0
# Note: MCP SDK - using OpenAI function calling as MCP protocol
# Tools are registered as OpenAI functions

# Rate Limiting
slowapi==0.1.9
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
aiosqlite==0.19.0  # For in-memory test database

# Type Checking and Linting
mypy==1.8.0
ruff==0.1.11

# Logging
python-json-logger==2.0.7
